Add generic FilterAndProject, Aggreation and HashAggregation

+++ presto-main/src/main/java/com/facebook/presto/nblock/uncompressed/UncompressedBlock.java
@@ -62,9 +62,9 @@ public class UncompressedBlock
         return rawRange;
     }
 
-    public int getCount()
+    public long getCount()
     {
-        return (int) (range.getEnd() - range.getStart() + 1);
+        return range.getEnd() - range.getStart() + 1;
     }
 
     @Override
